Apply suggestions of reviewer
